@@ -4,7 +4,6 @@ import (
 	"encoding/json"
 	"flag"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"strconv"
 
@@ -20,7 +19,7 @@ func JSON(f *string) Source {
 			return nil
 		}
 
-		j, err := ioutil.ReadFile(*f)
+		j, err := os.ReadFile(*f)
 		if err != nil {
 			return err
 		}
@@ -42,7 +41,7 @@ func dJSON(y []byte) Source {
 // using https://pkg.go.dev/github.com/drone/envsubst?tab=overview
 func YAML(f string, expandEnvVars bool) Source {
 	return func(dst Cloneable) error {
-		y, err := ioutil.ReadFile(f)
+		y, err := os.ReadFile(f)
 		if err != nil {
 			return err
 		}
@@ -74,17 +73,10 @@ func YAMLFlag(args []string, name string) Source {
 		// parsing out the config file location.
 		dst.Clone().RegisterFlags(freshFlags)
 
-		usage := freshFlags.Usage
 		freshFlags.Usage = func() { /* don't do anything by default, we will print usage ourselves, but only when requested. */ }
 
-		err := freshFlags.Parse(args)
-		if err == flag.ErrHelp {
-			// print available parameters to stdout, so that users can grep/less it easily
-			freshFlags.SetOutput(os.Stdout)
-			usage()
-			os.Exit(2)
-		} else if err != nil {
-			fmt.Fprintln(freshFlags.Output(), "Run with -help to get list of available parameters")
+		if err := freshFlags.Parse(args); err != nil {
+			fmt.Fprintln(freshFlags.Output(), "Run with -help to get a list of available parameters")
 			os.Exit(2)
 		}
 
@@ -99,6 +91,5 @@ func YAMLFlag(args []string, name string) Source {
 		}
 
 		return YAML(f.Value.String(), expandEnv)(dst)
-
 	}
 }
