@@ -5,7 +5,6 @@ import (
 	"context"
 	"fmt"
 	"io/fs"
-	"io/ioutil"
 	"net/http"
 	"os"
 	"strings"
@@ -38,17 +37,17 @@ import (
 func TestCreateLocalDir(t *testing.T) {
 	dataPath := t.TempDir()
 	localFile := dataPath + "/local"
-	require.NoError(t, ioutil.WriteFile(localFile, []byte("d"), 0o644))
+	require.NoError(t, os.WriteFile(localFile, []byte("d"), 0o644))
 	_, err := New(context.Background(), Config{
 		DataPath:         dataPath,
 		MaxBlockDuration: 30 * time.Minute,
-	})
+	}, NoLimit)
 	require.Error(t, err)
 	require.NoError(t, os.Remove(localFile))
 	_, err = New(context.Background(), Config{
 		DataPath:         dataPath,
 		MaxBlockDuration: 30 * time.Minute,
-	})
+	}, NoLimit)
 	require.NoError(t, err)
 }
 
@@ -143,7 +142,7 @@ func TestMergeProfilesStacktraces(t *testing.T) {
 	db, err := New(context.Background(), Config{
 		DataPath:         testDir,
 		MaxBlockDuration: time.Duration(100000) * time.Minute, // we will manually flush
-	})
+	}, NoLimit)
 	require.NoError(t, err)
 	defer require.NoError(t, db.Close())
 
@@ -271,7 +270,7 @@ func TestMergeProfilesPprof(t *testing.T) {
 	db, err := New(context.Background(), Config{
 		DataPath:         testDir,
 		MaxBlockDuration: time.Duration(100000) * time.Minute, // we will manually flush
-	})
+	}, NoLimit)
 	require.NoError(t, err)
 	defer require.NoError(t, db.Close())
 
