@@ -143,6 +143,8 @@ type Head struct {
 	tables          []Table
 	delta           *deltaProfiles
 	pprofLabelCache labelCache
+
+	limiter TenantLimiter
 }
 
 const (
@@ -151,7 +153,8 @@ const (
 	defaultFolderMode = 0o755
 )
 
-func NewHead(phlarectx context.Context, cfg Config) (*Head, error) {
+func NewHead(phlarectx context.Context, cfg Config, limiter TenantLimiter) (*Head, error) {
+	// todo if tenantLimiter is nil ....
 	parquetConfig := *defaultParquetConfig
 	h := &Head{
 		logger:  phlarecontext.Logger(phlarectx),
@@ -166,6 +169,7 @@ func NewHead(phlarectx context.Context, cfg Config) (*Head, error) {
 		flushForcedTimer: time.NewTimer(cfg.MaxBlockDuration),
 
 		parquetConfig: &parquetConfig,
+		limiter:       limiter,
 	}
 	h.headPath = filepath.Join(cfg.DataPath, pathHead, h.meta.ULID.String())
 	h.localPath = filepath.Join(cfg.DataPath, pathLocal, h.meta.ULID.String())
@@ -309,8 +313,15 @@ func (h *Head) convertSamples(ctx context.Context, r *rewriter, in []*profilev1.
 }
 
 func (h *Head) Ingest(ctx context.Context, p *profilev1.Profile, id uuid.UUID, externalLabels ...*typesv1.LabelPair) error {
-	metricName := phlaremodel.Labels(externalLabels).Get(model.MetricNameLabel)
 	labels, seriesFingerprints := labelsForProfile(p, externalLabels...)
+
+	for i, fp := range seriesFingerprints {
+		if err := h.limiter.AllowProfile(fp, labels[i], p.TimeNanos); err != nil {
+			return err
+		}
+	}
+
+	metricName := phlaremodel.Labels(externalLabels).Get(model.MetricNameLabel)
 
 	// create a rewriter state
 	rewrites := &rewriter{}
